+++ test/subdataframe_mutation.jl
@@ -1365,7 +1365,11 @@ end
     df = DataFrame(a=1:3)
     sdf = @view df[[3, 2], :]
     sdf.a .= 12.0
-    @test eltype(sdf.a) === Int
+    if isdefined(Base, :dotgetproperty)
+        @test eltype(sdf.a) === Float64
+    else
+        @test eltype(sdf.a) === Int
+    end
     @test df ≅ DataFrame(a=[1, 12, 12])
 
     if VERSION >= v"1.7"
@@ -1379,7 +1383,11 @@ end
     sdf = @view df[[3, 2], 1:1]
     @test_throws ArgumentError sdf.c = [5, 6]
     sdf.a .= 12.0
-    @test eltype(sdf.a) === Int
+    if isdefined(Base, :dotgetproperty)
+        @test eltype(sdf.a) === Float64
+    else
+        @test eltype(sdf.a) === Int
+    end
     @test df ≅ DataFrame(a=[1, 12, 12])
 end